*Pointのvaluetypeをdoubleからintへ変更 *固定小数点数の小数部を16ビットへ縮小 *Planeを強化

+++ isom_point.hpp
@@ -10,7 +10,7 @@
 struct
 Point
 {
-  using value_type = double;
+  using value_type = int32_t;
 
   value_type  x;
   value_type  y;
